+++ tests/unit-global/esnext.iterator.to-async.js
@@ -1,4 +1,5 @@
 import { STRICT } from '../helpers/constants.js';
+import ITERATOR from 'core-js-pure/es/symbol/iterator';
 
 QUnit.test('Iterator#toAsync', assert => {
   const { toAsync } = Iterator.prototype;
@@ -14,10 +15,29 @@ QUnit.test('Iterator#toAsync', assert => {
     assert.throws(() => toAsync.call(null), TypeError);
   }
 
+  const closableIterator = {
+    closed: false,
+    [ITERATOR]() { return this; },
+    next() {
+      return { value: Promise.reject(42), done: false };
+    },
+    return() {
+      this.closed = true;
+      return { value: undefined, done: true };
+    },
+  };
+
   return [1, 2, 3].values().toAsync().map(it => Promise.resolve(it)).toArray().then(it => {
     assert.arrayEqual(it, [1, 2, 3]);
     return new Set([1, 2, 3]).values().toAsync().map(el => Promise.resolve(el)).toArray();
   }).then(it => {
     assert.arrayEqual(it, [1, 2, 3]);
+  }).then(() => {
+    return Iterator.from(closableIterator).toAsync().toArray();
+  }).then(() => {
+    assert.avoid();
+  }, error => {
+    assert.same(error, 42, 'rejection on a callback error');
+    assert.true(closableIterator.closed, 'doesn\'t close sync iterator on promise rejection');
   });
 });