+++ tests/unit-pure/esnext.array.from-async.js
@@ -1,8 +1,9 @@
-import { createAsyncIterable, createIterable } from '../helpers/helpers.js';
+import { createAsyncIterable, createIterable/* , createIterator */ } from '../helpers/helpers.js';
 import { STRICT_THIS } from '../helpers/constants.js';
 
 import Promise from 'core-js-pure/es/promise';
 import fromAsync from 'core-js-pure/actual/array/from-async';
+// import Iterator from 'core-js-pure/actual/iterator';
 
 QUnit.test('Array.fromAsync', assert => {
   assert.isFunction(fromAsync);
@@ -19,6 +20,10 @@ QUnit.test('Array.fromAsync', assert => {
   assert.same(counter, 1, 'proper number of constructor calling');
 
   function C() { /* empty */ }
+
+  // const closableIterator = createIterator([1], {
+  //   return() { this.closed = true; return { value: undefined, done: true }; },
+  // });
 
   return fromAsync(createAsyncIterable([1, 2, 3]), it => it ** 2).then(it => {
     assert.arrayEqual(it, [1, 4, 9], 'async iterable and mapfn');
@@ -86,4 +91,13 @@ QUnit.test('Array.fromAsync', assert => {
   }, error => {
     assert.true(error instanceof TypeError);
   });
+  /* Tests are temporarily disabled due to the lack of proper async feature detection in native implementations.
+  .then(() => {
+    return fromAsync(Iterator.from(closableIterator), () => { throw 42; });
+  }).then(() => {
+    assert.avoid();
+  }, error => {
+    assert.same(error, 42, 'rejection on a callback error');
+    assert.true(closableIterator.closed, 'doesn\'t close sync iterator on promise rejection');
+  }); */
 });